refactor: fix type validation errors and enhance code stability
